Move Actors to separate files
