+++ chain/network/src/lib.rs
@@ -1,6 +1,7 @@
 pub use peer::{EPOCH_SYNC_PEER_TIMEOUT_MS, EPOCH_SYNC_REQUEST_TIMEOUT_MS};
-pub use peer_manager::{
-    PeerManagerActor, RoutingTableActor, RoutingTableMessages, RoutingTableMessagesResponse,
+pub use peer_manager::PeerManagerActor;
+pub use routing_table_actor::{
+    RoutingTableActor, RoutingTableMessages, RoutingTableMessagesResponse,
 };
 pub use types::{
     FullPeerInfo, NetworkAdapter, NetworkClientMessages, NetworkClientResponses, NetworkConfig,
@@ -9,6 +10,7 @@ pub use types::{
 
 mod cache;
 mod codec;
+mod edge_verifier;
 mod ibf;
 pub mod ibf_peer_set;
 pub mod ibf_set;
@@ -18,6 +20,7 @@ mod peer_manager;
 pub mod peer_store;
 mod rate_counter;
 pub mod routing;
+mod routing_table_actor;
 pub mod test_utils;
 pub mod types;
 pub mod utils;
